+++ Partie 2/Nombre.java
@@ -1,21 +1,21 @@
 public class Nombre extends Expression {
     // Attribut pour stocker la valeur du nombre
-    private int valeurNombre;
+    private double valeurNombre;
 
     // Constructeur pour initialiser la valeur du nombre
-    public Nombre(int valeurNombre) {
+    public Nombre(double valeurNombre) {
         this.valeurNombre = valeurNombre;
     }
 
     // Méthode pour retourner la valeur du nombre
     @Override
-    public int valeur() {
+    public double valeur() {
         return valeurNombre;
     }
 
     // Méthode pour retourner la représentation en chaîne de la valeur du nombre
     @Override
     public String toString() {
-        return Integer.toString(valeurNombre);
+        return Double.toString(valeurNombre);
     }
 }
